requirements/base.txt
Django==3.0.7
django-environ==0.4.5
psycopg2==2.8.5
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
django-allauth==0.41.0
django-avatar==5.0.0
six==1.14.0
